requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
slack==0.0.2
slack-sdk==3.21.3
slackclient==2.9.4
yarl==1.9.2
